[AIRFLOW-6721] Move TestHiveEnvironment to test module
Movee the TestHiveEnvironment class and setUp method out to the hive
test module so that it can be used in other hive tests.
